ref: eliminate tests.sentry.plugins.testutils by moving code to modules using it
